feat: break audioplaylist into reusable components
